서범수_회원가입시문자인증진행중_20191218

See merge request whyj97/hagong!221

+++ src/hagong/academy/mngClass/mngCS/controller/InsertClassStudentServlet.java
@@ -31,24 +31,35 @@ public class InsertClassStudentServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String stuArr = request.getParameter("stuArr");
 		String stuAll = request.getParameter("stuAll");
-		
-		String[] stuAll2 = stuAll.split(",");
-		String[] stuArr2 = stuArr.split(",");
-		int[] stuNo = new int[stuArr2.length];
-		int[] stuAllNo = new int[stuAll2.length];
-		
-		for(int i = 0; i < stuArr2.length; i++) {
-			stuNo[i] = Integer.parseInt(stuArr2[i]);
-		}
-		for(int i = 0; i < stuAll2.length; i++) {
-			stuAllNo[i] = Integer.parseInt(stuAll2[i]);
-		}
 		int clsNo = Integer.parseInt(request.getParameter("clsNo"));
+		int resultInsert = 0;
+		int resultDelete = 0;
+		
+		if(!stuArr.equals("first")) {
+			String[] stuArr2 = stuArr.split(",");
+			int[] stuNo = new int[stuArr2.length];
+			for(int i = 0; i < stuArr2.length; i++) {
+				stuNo[i] = Integer.parseInt(stuArr2[i]);
+			}
+			resultInsert = new CSService().insertCS(clsNo, stuNo);
+			if(resultInsert > 0) {
+				resultDelete = 1;				
+			}
+		}
+		if(!stuAll.equals("second")) {
+			String[] stuAll2 = stuAll.split(",");
+			int[] stuAllNo = new int[stuAll2.length];
+			
+			for(int i = 0; i < stuAll2.length; i++) {
+				stuAllNo[i] = Integer.parseInt(stuAll2[i]);
+			}			
+			resultDelete = new CSService().deleteCS(clsNo, stuAllNo);
+			if(resultDelete > 0) {
+				resultInsert = 1;				
+			}
+		}
 		
 		
-		
-		int resultInsert = new CSService().insertCS(clsNo, stuNo);
-		int resultDelete = new CSService().deleteCS(clsNo, stuAllNo);
 		
 		int result = 0;
